III-4481 - Make seperate component for PriceInformation on step 5
