+++ src/ui/theme.ts
@@ -319,9 +319,7 @@ const theme = {
       hoverTabBackgroundColor: colors.grey3,
       borderRadius: 0,
     },
-    additionalInformationStep: {
-      borderColor: colors.grey3,
-    },
+
     pictureUploadBox: {
       backgroundColor: colors.white,
       borderColor: colors.grey2,
@@ -332,7 +330,7 @@ const theme = {
       mainImageBorderColor: colors.blue2,
       thumbnailBorderColor: colors.grey6,
     },
-    priceInfoModal: {
+    priceInformation: {
       borderColor: colors.grey3,
     },
     videoUploadBox: {
